GameProvider saperation of concerns , encapsulation context
