--- src/context/cellsContext.js
@@ -1,3 +1,0 @@
-import { createContext } from "react";
-
-export const cellsContext = createContext(null);
